Replace manual plumbing with Analyzer typeclass
This reduces duplication, but the main impetus of this change is the
anticipation of separating analyzeTerm and analyzeProperty to use two
different monad stacks, for two reasons:

1. as we share more implementation between analysis of Term and Prop,
and perhaps add support for reading tables from the property language,
we want to ensure that analyzeProperty has strictly read-only access to
AnalyzeState.
2. property analysis should have access to the symbolic "result" from
term analysis. Currently this result is discarded.