Implement quantifier floating.
Because SBV assumes inputs are in prenex normal form, we need to do this
prenex conversion before handing off quantified properties.

There's one significant difficulty here -- the fact that quantifier
names can collide. Example:

`(forall a. a > 5 => a > 4) /\ (forall a. a \/ ~a)`

Naively syntactically lifting results in

`forall a. forall a. (a > 5 => a > 4) /\ (a \/ ~a)`

which is nonsense, since the two `a`s aren't even the same type.

Our chosen solution is to generate a unique id for each variable,
resulting in

`forall a0. forall a1. (a0 > 5 => a0 > 4) /\ (a1 \/ ~a1)`