Improve property parsing / checking.
The biggest improvement here is that we no longer have to specify
`int-column-conserves` / `dec-column-conserves` -- we can infer the type
and instead just specify `column-conserves`.

This change also improves handling of string-like identifiers, so
`"foo"` and `'foo` are handled uniformly.

We remove the `RowKey` `newtype`, so row keys are now treated
(correctly) as strings.

Lastly, parsing is simplified and more extensible.

Fixes #85