Introduce Model
This is just the first step of a few to integrate `Model`. Though we can
"show" all `Model` fields, we're not yet populating them; only args so
far. "Reads" and "auths" will probably come from a `WriterT` in the
translation step. Once this is complete, we will "tag" reads and auths
during analysis by logging equality constraints.