Only emit subpath, not root, path start
Only emitting subpath starts, and simply constraining the root path to
true in tag allocation seems slightly less awkward than plumbing the
root path `TagId` to analysis simply so that it can do the same thing,
and emitting an additional path-start at the beginning of the program.

The main driver of this decision is that we can easily/naturally put
path `TagId`s for each branch on an `IfThenElse` constructor, but
there's no natural place to do the same for the singleton start of a
program.

Tagging subpath starts during analysis will land shortly.