Fix cell and column deltas for inserts.
Previously it was undetermined what the value of a numeric cell was
before written for the first time. This should sound nonsensical, so in
more detail, consider the following line:

`(insert accounts address { "balance": 0.0 })`

We intuitively understand that the cell / column delta is zero, but
analysis would report:

```
Program trace:
  entering function create-account with argument
    address = ""

    write { balance: 0.0 } to key "" succeeds
    returning with "Write succeeded"
```

We track whether a row exists (which it must not before the insert), so
the model would simultaneously assert that the row did not previously
exist, but its previous balance was 1.

This is clearly nonsense, and the type of thing we'd like to avoid
longer term, perhaps by giving the system a stronger ontological
understanding of the existence of a row. For now, I believe only int /
decimal, row / cell deltas are affected, and all should be fixed by this
change.

Tested by
* Verified verified-accounts successfully
* Added "conserves-mass.integer.insert" test
* Added "conserves-mass.decimal.insert" test