Implement remote verification via REST API
Unfortunately in order to use XHR while using `reflex-platform` but
without taking on `reflex` as a dependency, we had no choice but to
copy/adapt the XHR implementation from `ghcjs-base`. We can't use the
XHR implementation from `ghcjs-base` directly because `reflex-platform`
patches `ghcjs-base` to remove the XHR module to avoid a circular
dependency involving `text`.

We had to butcher this XHR implementation a bit, because outside of
`ghcjs-base`, we don't have access to some of their internal modules.