Use a timeout of one second for all z3 queries.
Motivation:
Due to https://github.com/Z3Prover/z3/issues/2060, our CI was recently
failing for a couple months in a way that made it very hard to debug (we
just saw no output on travis' logs). This timeout will cause any future
z3 hanging to be diagnosable.

We've also previously considered adding a timeout for user queries
(which are more likely to take a long time). With this change queries
that take longer than a second will show a warning on the line with the
property / invariant being checked.

In the future we may want the timeout to be configurable.