Introduced CommandValue
so we can properly differentiate between `CommandError` and
`CommandSuccess`.

Previously on the client side one would have needed to try to parse an
error catch that parsing error and then try to parse success. With this
change, the client can simply parse `CommandValue` and get the result
(either success or error). `CommandSuccess` is now no longer a type of
its own, but a constructor of `CommandValue`. I kept `CommandError`, so
we don't have partial accessor functions for `CommandValue`.

What I am not sure about is:

1. Should we really return `CommandValue Value` from /local or should it be the more
type safe: `CommandValue (Term Name)`? For now I kept it as it was
before, which is `CommandValue Value`.
2. If we are going to return `Term Name`, I think `CommandValue` no longer
needs to be a type constructor, but could be a simple type isomorph
to `CommandValue (Term Name)`.
3. Is it really necessary to have `_crResult` to be of type `Value`,
can't it be `CommandValue` directly? This would make things more typesafe and the
REST API a bit more approachable.