Fix type error when binding only partial objects.
The error occurred in this example:

    (defschema account
      name:string
      balance:integer)

    (defun get-balance:integer (accountId:string)
        (with-default-read accounts-table accountId
          { "balance" : 0 }
          { "balance" := balance }
          balance
        )
    )

Note that `{ "balance" : 0 }` is not really of type `account`, since it
doesn't have a name field. The analysis system previously couldn't
handle this kind of ad-hoc typing. With this change we can now handle
this case, by checking that the default object and the binding type
match.

This is a slightly stronger restriction than we strictly require -- the
binding type could actually be smaller than the provided default, but
that seems generally indicative of a mistake. We do the more
conservative check for now.

This change affects all reads, changing the result type to only reflect
the fields which are actually read, though it should really only matter
in this case.

Tested via `cabal test` and running `examples/cp/cp.repl`.