Ensure we never property check lists longer than 10.
Previously, we could generate intermediate computations using lists of
length longer than 10, since we only checked if the final result was
longer than 10. Example of a test that actually failed:

    (enforce
      (=
        (make-list 10 0)
        (drop 1 (reverse (make-list 11 0)))
        ""))

The new method of generating lists only generates valid lists of length
<= 10. We do this by only allowing integer literals as the length
argumenent to `make-list`. The downside is that we rule out a bunch of
valid programs.

Ways lists are generated:
* `make-list`: we only `make-list` lists of a valid length
* `reverse` / `sort`: these preserve length
* `concat`: for each sublist we halve the allowed length
* `drop` / `take`: these always return a list no longer than their
  argument.