Detect and report vacuous properties
This is important for minimizing user confusion and surprise. Because
`property` assumes transaction success, when a function cannot
successfully run to completion, it is possible to "validate" any
`property` in this setting.

Consider some logical formula that corresponds to the Pact expression
`(property some_property)`, where `some_property` can be anything:

    tx_success => some_property

This is equivalent to:

    ¬tx_success Λ some_property

If `tx_success` is always false, then this expression is *always* true, no
matter the value of `some_property`.

We now detect such a condition where it is not possible for this
function to succeed (if called as a top-level function), and we produce
an error notifying the user that the property is moot.