Fix column delta calculation. (#799)
We were using 0 for the previous value only if the write type was an
insert. In the failing test case the write type was a... write. This
means, we were using the previous value of the cell, which is undefined
/ unbound. z3 took this to mean it could be anything, leading to
nonsensical results. The fix is simply to use 0 if the row did not
previously exist.