Error message fix
This fixes replay. We check in evalCap if a cap's arguments have the right type.
We used to check in installSigCap too, which would be called before evalCap.
I recognized while implementing verifiers that this check was unnecessary and
deleted it, but we need it to remain, because otherwise the call stack from
the typechecking error changes and breaks replay.